+++ lib/app/modules/order_stock_adjustment_register/domain/entity/order_stock_adjustment_register_items_entity.dart
@@ -7,7 +7,7 @@ class OrderStockAdjustmentRegisterItemsEntity {
   int tbProductId;
   String description;
   String nameStockList;
-  int unitValue;
+  double unitValue;
   double quantity;
 
   OrderStockAdjustmentRegisterItemsEntity({
